Remove `Codelist.definition_csv_data_for_download`.
This model imports from `presenters.py`. This means the presenters can't import
from the model they are presenting, without a circular import, which seems
backwards. I want to make such an import in a later commit for a `Prefetch`.

This is only used in one place so it doesn't see like it needs factoring out
into shared code, really. Let's just call it part of the view presenting the
data.