Use `Prefetch` for `SearchResult`.
I found that upgrading to Django 5.2 caused the query used with the nested
`select_related` exploded into a very long series of `OR` clauses that exceeded
database limits. Using `prefetch_related` is more natural based on the analysis
below and resolves that issue with upgrading to 5.2.

More info below and in the issue description (more or less the same).

Analysis
--------

Problem
-------

Problem `QuerySet`:

```
for search in clv.searches.prefetch_related(
    "results", "results__code_obj"
).order_by("code", "term"):
```

The problematic queries are large so to analyse them I wrote them and their
parameters to file by editing site-`packages/django/db/backends/sqlite3/base.py
SQLiteCursorWrapper.execute()`:

```
query = self.convert_query(query, param_names=param_names)
with open('query-51.sql', 'a') as ff:
    ff.write(f'{query} % {params}\n')
```

The relevant models are structured like this:

```
--------        --------------        ---------
|Search| 1 -- M |SearchResult| M -- 1 |CodeObj|
--------        --------------        ---------
```

`Search` is a record of something the user typed in the search bar.
`SearchResults` are the hits for that search in the relevant coding system.
`CodeObj` represents the choice to include a particular code in this codelist,
or not.

One Search can yield many `SearchResults`, which have only one `CodeObj` each
(but each `CodeObj` can be shared between many `SearchResults`, as similar
searches may yield the same code, and we only want the user to need to decide
on each code once). This is effectively a many-to-many relationship between
`Search` and `CodeObj`.

Aside: I think we would be better off replacing `SearchResult` with an explicit
`ManyToMany` field, to avoid issues like this. Or simply not recording that
relationship in the DB at all, as it can be reconstructed when accessed and
cached in memory.  Out of scope for this ticket.

The `QuerySet` runs three queries. One to fetch Searches matching the
particular `CodelistVersion`, one to fetch `SearchResults` matching that
search, and one to fetch the `CodeObj` for each search result. The queries look
like (simplified):

```
SELECT {FIELDS} FROM "search" WHERE ("version_id" = 123)
SELECT {FIELDS} FROM "searchresult" WHERE ("search_id" IN (12, 34, 56, 78))
SELECT {FIELDS} FROM "codeobj" WHERE ("id" = 523 OR "id" = 556 OR OR "id" = 589 OR ...)
```

The third query uses the literal values of the `SearchResult` identifiers from
the second search. When there are many search results, this can lead to a very
long query which can exceed the database's limits, causing the query to fail
and an exception to propagate.

I checked the relevant limits locally:

```
~/opencodelists$ sqlite3 db.sqlite3
SQLite version 3.37.2 2022-01-06 13:25:41
Enter ".help" for usage hints.
sqlite> PRAGMA compile_options;
...
MAX_EXPR_DEPTH=1000
MAX_SQL_LENGTH=1000000000
MAX_VARIABLE_NUMBER=250000
```

Some codelist versions have more than 1000 search results, so the third query breaches the expression depth limit.

Difference in Django 5.1 to 5.2
-------------------------------

In Django 5.1, the query uses an `IN` clause `("id" IN (1, 4, 5, ...)`.
Apparently this did not ever breach the database's variable limit for an `IN`
clause (250000).

In Django 5.2, the query instead uses `OR` concatenated clauses instead `WHERE
("id" = 1 OR "id" = 4 OR OR "id" = 5 OR ...)`. This did breach the database's
nesting level limit for many codelists (each `OR` clause increases the level of
nesting of the logic expression by one). Leading to the failure and exception
for codelists with many search results.

I don't understand exactly why Django behaviour has changed here with the exact
some database and database version etc.

There were some changes to how prefetches are handles in 5.2, as Lucy noted
leading to [at least one
regression](https://code.djangoproject.com/ticket/36432) (though that was not
directly related to our issue). But I haven't found the specific change that
caused this in the release notes and haven't gone looking in the code -- I
think fixing the prefetch avoids the issue and is the right thing to do anyway.

We can fix the prefetching to use an `INNER JOIN` in the second query rather
than a long list of IDs in a third query by changing the call slightly:

```
for search in clv.searches.prefetch_related(Prefetch(
    "results",
    queryset=SearchResult.objects.select_related("code_obj"),
)).order_by("code", "term"):
```

This leads to two queries like (simplified):

```
SELECT {FIELDS} FROM "search" WHERE ("version_id" = 123)
SELECT {FIELDS} FROM "searchresult" INNER JOIN codeobj ON "searchresult.code_obj"_id = "codeobj.id" WHERE ("search_id" IN (12, 34, 56, 78))
```

This is a much smaller query and doesn't breach limits unless the number of
searches is higher than the database's variable limit of 250000 (see above).

We can use `select_related` because there is at most (or in this case, exactly)
one `CodeObj` per `SearchResult`. In general, it is better to use
`select_related` when there can only be one related object, which is the case
here. `prefetch_related` is for `ManyToMany` relationships or reverse
`ForeignKeys` (like `Search` to `SearchResult`).

I haven't thought in detail about the memory/speed/query size trade-off in
great detail as this change seems like clearly the right thing to do. If there
were further problems with very large queries we could consider other
strategies.

Testing
-------

I found the following codelist that exhibits the issue on the first page of the index with a prod mirror:

http://localhost:7000/codelist/opensafely/acute-respiratory-illness-primary-care/30f1c6ca/

And from looking in #tech-incidents:

http://localhost:7000/codelist/opensafely/weight-snomed/5459abc6/

These loads fine in 5.1 or with the change in 5.2 and throw the exception in
5.2 without the change. Performance did not seem to degrade, actually
marginally improved (compared to 5.1) according to debug toolbar

5.1:
```
3 queries that run in 60ms
Whole page 750-850ms CPu time
Whole page 28 queries in 40-60ms
```

5.2
```
Two queries that run in .63ms
Whole page 700-800ms CPu time
Whole page 27 queries in 30-40ms

After the change all automated tests pass.

In theory there might be other pages with queries that breach some of the
limits given some state of the database, but this seems very hard to
exhaustively check for in principle. In practice, we didn't observe any other
issues around the incident so I expect this is the only issue.

We could consider a fixture / test with a very large number of searches and
results for testing the presenter/view and elsewhere. We could also consider
somehow testing views on EVERY codelist in a prod mirror, or a
representative/worst sample. Out of scope for this ticket.
```