Update visible_versions to include current version
If the user is currently on an "Under Review" codelist, that codelist
doesn't show in the versions list. That means the "Created" date is
not visible.

This updates the visible_versions code to check for a matching ID,
and checks if the user has permissions to see other versions using an
OR statement.