Natively compare Hierarchies instead of their json
Hierarchy uses sets to contain collections of codes.
When serialised then deserialised from json these become lists.
When comparing deserialised cached hierarchies either with other
deserialised cached hierarchies or freshly-constructed hierarchies the
order of these lists is not deterministic.

The only place that hierarchies are compared currently is in release
compatibility checking.
Rather than constructing Hierarchies, serialising them to json and
comparing that json (excluding the bits we're not interested in),
we can override the equality method for Hierarchy to take account of
these comparison requirements and just natively compare Hierarchies.