Migration to remove `Codelist` with no `Handle`.
Per https://github.com/opensafely-core/opencodelists/pull/2919/ and
https://github.com/opensafely-core/opencodelists/issues/2893 `Codelist` should
not be able to exist without `Handle`. We found an example, let's clean it up.

Let's do this in a migration in a new maintenance app, so that the fix is
auditable, reviewable, repeatable, testable, avoids the possibility of manual
error, and fixes up development environment databases as well. This also
provides an example if we need to do this or other maintenance again. And it
avoids cluttering the feature apps and getting squashed.

The migration is idempotent and has a noop reversal.

Manual testing with the latest sanitised backup (with judicious snipping of
noise throughout):

Before the migration migrations state:
```
~/opencodelists$ just manage showmigrations
...
maintenance
 [ ] 0001_delete_codelists_without_handles
```

Before state in the Django shell,  1 such codelist exists:
```
>>> len(Codelist.objects.all())
10419
>>> no_handles = Codelist.objects.filter(handles__isnull=True)
>>> len(no_handles)
1
```

Applying the migration:
```
just manage migrate maintenance 0001
Running migrations:
	Applying maintenance.0001_delete_codelists_without_handles...
0001: Deleted 1 codelists without handles

just manage showmigrations
...
maintenance
 [X] 0001_delete_codelists_without_handles
```

After state in the Django shell, no such codelist exists and only 1 codelist has been deleted:
```
len(Codelist.objects.all())
10418
>>> no_handles = Codelist.objects.filter(handles__isnull=True)
>>> len(no_handles)
0
```

Re-applying a very similr migration based on the same file and depending on it
(not committed), to to test the `except` branch:

```
just manage migrate maintenance 0002
0002: No codelists without handles to delete
```

This safely demonstrates the migration works as expected on near-production
data and also in databases without the faulty codelists. This isn't part of
ongoing application code so it does not need automated testing.