Update release history compatibility
PR #2911 fixed the calculation of coding system release compatibility
for codelist versions.

It did not correct the incorrectly calculated historical compatibility.

This commit introduces a way of fixing that history via a Django
migration.

In order to achieve this, the dynamically-set coding system database
connections must be in place.
The function that does this- `update_coding_system_database_connections`
checks to see if any migrations are pending and does not update the
connections if so - meaning it is unusable in a migration.
Therefore, this commit also adds an override mode to this function such
that it can be used in a migration.

The migration then iterates through all releases of all coding systems
(in release date order) and calls the
`update_codelist_version_compatibility` function that would normally be
used in a coding system release import.