Fix link html parsing for Sept 2025 release
When the September release was moved to the historical releases section
with the release of the October release,
the HTML for its link was in a new and unexpected format:

<p>September<a href="/sites/default/files/2025-11/BNF%20Snomed%20Mapping%20data%2020251120.zip"> 2025 (ZIP file: 18.77MB)</a></p>

<p><a href="/sites/default/files/2025-10/BNF%20Snomed%20Mapping%20data%2020251017.zip">August 2025 (ZIP file: 18.2MB)</a></p>

The previous code assumed that if the link text didn't fit the
"month YYYY" pattern then only the month part of the "valid from" date
was part of the link text and the year part was in the following element.
This was the only form of malformed historical release link we'd seen,
but this fails with this new case since the missing date part is in the
preceding element (and there is no following element).

This commit accounts for both these types of malformed link,
with an accompanying test.